Изменен тип поля accrual в AccrualCalculationRes

+++ internal/app/domain/orders.go
@@ -10,9 +10,9 @@ type OrderDTO struct {
 }
 
 type AccrualCalculationRes struct {
-	Order   string `json:"order"`
-	Status  string `json:"status"`
-	Accrual int    `json:"accrual"`
+	Order   string  `json:"order"`
+	Status  string  `json:"status"`
+	Accrual float32 `json:"accrual"`
 }
 
 const OrderProcessedStatus = "PROCESSED"
